Start project with basic setup and one block
Initialized the project for Minecraft 1.16.1 with Forge 32.0.98. There
are no proxies and @SubscribeEvents because I want to stick to the latest
Forge recommendations which are using DeferredRegistries and listeners.
The basic block named 'Diamond Generator' is equipped with BlockState,
model, texture, loot table and recipe.
Supported languages are British English and American English.