Redesign Diamond Generator
Model was changed to reflect the cubic structure of the generator and
correctly display new textures made by Kiszczame. The new textures are
basic.png for boilerplate, frame.png for connections, vent.png which
represents ventilation and diamond_generator.png which represents
the front face of the generator. DiamondGenerator was also changed,
namely getShape method was removed, as it does not need custom shape
anymore.