Add Tungsten Pickaxe and clear the code
Added the Tungsten Pickaxe item, registered it, assigned it to the mod's
creative tab, added the translation and the recipe.
Renamed registry classes to match a naming pattern of 'Type' + 'Name',
added said 'Type' to the Diamond Generator class.
Removed us lang file as it is useless now and only creates clutter.