Add GUI for Diamond Generator
Added a Container class for Client-Server cooperation and registered it
in a new RegistryContainer class. Added a ServerSide-only code respon-
sible for opening the GUI inside the Block class and ClientSide-only
new class Screen for rendering that GUI.
When BlockDiamondGenerator#onBlockActivated is called there is a test
which checks if the world is remote and if it is on the server side
it calls NetworkHooks.openGui(). One of the parameters is a class
implementing INamedContainerProvider, in our case it is
TileDiamondGenrator for simplicity. This guarantees the presence of the
createMenu method with returns the new ContainerDiamondGenerator
instance. Container object hold slots ocjects which help both
Client and ServerSide cooperate with ItemStacks.
Container is also a parameter of the ScreenDiamondGenerator, which
renders the GUI ClientSide. Screen factory is registered in the ModSetup
class and consist of tuple of Container and Screen supplier.