Add new WeirdMob entity
Updated forge for 108 ver. Added new Entity and Registered it in the new
RegistryEntity. Created a model for it and a renderer too. Registering
renderer happens only client-side. Added a Spawn Egg for this new entity
which unfortunatelly isn't currently supported by Forge.
Practically used the Cadiboo way of adding it. The egg also was registe-
red. Of course added localization for the new objects.
Minor addition: added a more advanced tooltip for DiamondGenerator, it
shows now Energy stored and diamonds left.